:bug: replace delius event id Int with Long

+++ src/main/kotlin/uk/gov/justice/digital/assessments/api/CreateAssessmentDto.kt
@@ -5,7 +5,7 @@ import uk.gov.justice.digital.assessments.jpa.entities.AssessmentType
 
 data class CreateAssessmentDto(
   @Schema(description = "Delius Event ID", example = "1234")
-  val deliusEventId: Int? = null,
+  val deliusEventId: Long? = null,
 
   @Schema(description = "Offender CRN", example = "CRN1")
   val crn: String? = null,
